file, sprite: move process_image to sprite module
